test: 게시글을 생성할 때, postRepository가 모킹되어있으면 응답의 게시글 id값이 null이 나오는 이슈로 테스트 클래스 분리
